+++ docs/showcase/permission-matrix.tsx
@@ -173,7 +173,6 @@ export default function Demo() {
                   value={compareA}
                   onValueChange={setCompareA}
                   options={roleOptions}
-                  clearable={false}
                   className="w-40"
                 />
               </FormField>
@@ -182,7 +181,6 @@ export default function Demo() {
                   value={compareB}
                   onValueChange={setCompareB}
                   options={roleOptions}
-                  clearable={false}
                   className="w-40"
                 />
               </FormField>